fix: save correct curation.json to results

+++ code/run_capsule.py
@@ -345,8 +345,8 @@ if __name__ == "__main__":
                         analyzer.set_sorting_property("decoder_probability", values, save=True)
         _ = analyzer.sorting.save(folder=curated_results_folder / recording_name)
 
-        curation_json_files = [p for p in curated_folder.iterdir() if p.name.startswith("curation_")]
-        for curation_json_file in curation_json_files:
+        curation_json_file = curated_folder / f"curation_{recording_name}.json"
+        if curation_json_file.is_file():
             shutil.copyfile(curation_json_file, curated_results_folder / recording_name / "curation.json")
 
         # If the collect results runs in a pipeline, we need to further modify the mappings of the preprocessed recording in the analyzer.
